server: remove redundant duplicate initPerformance
